visualization 1개 표 저장, make gym_util.py, stack_visualization_data func util.py로 이동
